requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis and Caching
redis==5.0.1
celery==5.3.4

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Image Processing
Pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
diffusers==0.24.0
transformers==4.35.2
accelerate==0.24.1

# Validation and Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Environment
python-dotenv==1.0.0

# File Storage
boto3==1.34.0  # For S3 compatibility
minio==7.2.0   # For MinIO storage

# Additional utilities
click==8.1.7
rich==13.7.0
typer==0.9.0
